# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "radbrain API",
    "description": "Provenance-first, tenant-isolated radiology study platform.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/preview/onboarding": {
      "post": {
        "tags": [
          "preview: non-release"
        ],
        "summary": "Onboard",
        "operationId": "onboard_v1_preview_onboarding_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewOnboardingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewPlanResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/preview/plan": {
      "get": {
        "tags": [
          "preview: non-release"
        ],
        "summary": "Plan",
        "operationId": "plan_v1_preview_plan_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewPlanResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/preview/plan/replan": {
      "post": {
        "tags": [
          "preview: non-release"
        ],
        "summary": "Regenerate Plan",
        "operationId": "regenerate_plan_v1_preview_plan_replan_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewPlanResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/preview/today": {
      "get": {
        "tags": [
          "preview: non-release"
        ],
        "summary": "Today Bundle",
        "operationId": "today_bundle_v1_preview_today_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewTodayResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/preview/cards/due": {
      "get": {
        "tags": [
          "preview: non-release"
        ],
        "summary": "Cards Due",
        "operationId": "cards_due_v1_preview_cards_due_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PreviewCardResponse"
                  },
                  "title": "Response Cards Due V1 Preview Cards Due Get"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/preview/cards/{card_id}/review": {
      "post": {
        "tags": [
          "preview: non-release"
        ],
        "summary": "Review",
        "operationId": "review_v1_preview_cards__card_id__review_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Card Id"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewCardResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/preview/mastery": {
      "get": {
        "tags": [
          "preview: non-release"
        ],
        "summary": "Mastery View",
        "operationId": "mastery_view_v1_preview_mastery_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "x-tenant-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "x-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewMasteryResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Citation": {
        "properties": {
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "page_no": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page No"
          },
          "block_id": {
            "type": "string",
            "title": "Block Id"
          },
          "bbox": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Bbox"
          },
          "figure_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Figure Id"
          }
        },
        "type": "object",
        "required": [
          "source_id",
          "page_no",
          "block_id",
          "bbox"
        ],
        "title": "Citation"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PreviewCardResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "citation": {
            "$ref": "#/components/schemas/Citation"
          },
          "due_at": {
            "type": "string",
            "format": "date-time",
            "title": "Due At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "prompt",
          "citation",
          "due_at"
        ],
        "title": "PreviewCardResponse"
      },
...
      "PreviewMasteryResponse": {
        "properties": {
          "mode": {
            "type": "string",
            "const": "preview",
            "title": "Mode",
            "default": "preview"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "nodes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Nodes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version",
          "method",
          "nodes"
        ],
        "title": "PreviewMasteryResponse"
      },
      "PreviewOnboardingRequest": {
        "properties": {
          "exam_date": {
            "type": "string",
            "format": "date",
            "title": "Exam Date"
          },
          "hours_per_week": {
            "type": "integer",
            "maximum": 40.0,
            "minimum": 1.0,
            "title": "Hours Per Week"
          },
          "session_minutes": {
            "type": "integer",
            "enum": [
              30,
              60,
              90
            ],
            "title": "Session Minutes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "exam_date",
          "hours_per_week",
          "session_minutes"
        ],
        "title": "PreviewOnboardingRequest"
      },
...
      "PreviewPlanResponse": {
        "properties": {
          "mode": {
            "type": "string",
            "const": "preview",
            "title": "Mode",
            "default": "preview"
          },
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "plan_version": {
            "type": "integer",
            "title": "Plan Version"
          },
          "exam_date": {
            "type": "string",
            "title": "Exam Date"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "hours_per_week": {
            "type": "integer",
            "title": "Hours Per Week"
          },
          "session_minutes": {
            "type": "integer",
            "title": "Session Minutes"
          },
          "days_remaining": {
            "type": "integer",
            "title": "Days Remaining"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "priority": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Priority"
          },
          "baseline_status": {
            "type": "string",
            "title": "Baseline Status"
          },
          "weight_policy": {
            "type": "string",
            "title": "Weight Policy"
          },
          "notice": {
            "type": "string",
            "title": "Notice"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "plan_id",
          "plan_version",
          "exam_date",
          "timezone",
          "hours_per_week",
          "session_minutes",
          "days_remaining",
          "phase",
          "priority",
          "baseline_status",
          "weight_policy",
          "notice"
        ],
        "title": "PreviewPlanResponse"
      },
...
      "PreviewReviewRequest": {
        "properties": {
          "rating": {
            "type": "integer",
            "enum": [
              1,
              2,
              3,
              4
            ],
            "title": "Rating"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rating"
        ],
        "title": "PreviewReviewRequest"
      },
...
      "PreviewTodayResponse": {
        "properties": {
          "mode": {
            "type": "string",
            "const": "preview",
            "title": "Mode",
            "default": "preview"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "plan_version": {
            "type": "integer",
            "title": "Plan Version"
          },
          "days_remaining": {
            "type": "integer",
            "title": "Days Remaining"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "blocks": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Blocks"
          },
          "omitted_blocks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Omitted Blocks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "plan_version",
          "days_remaining",
          "phase",
          "blocks",
          "omitted_blocks"
        ],
        "title": "PreviewTodayResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}